feat(cast): implement ERC-7201 as `cast index-erc7201` (#8092)
