+++ crates/cast/bin/opts.rs
@@ -526,6 +526,16 @@ pub enum CastSubcommand {
         slot_number: String,
     },
 
+    /// Compute storage slots as specified by `ERC-7201: Namespaced Storage Layout`.
+    #[command(name = "index-erc7201", alias = "index-erc-7201", visible_aliases = &["index7201", "in7201"])]
+    IndexErc7201 {
+        /// The arbitrary identifier.
+        id: Option<String>,
+        /// The formula ID. Currently the only supported formula is `erc7201`.
+        #[arg(long, default_value = "erc7201")]
+        formula_id: String,
+    },
+
     /// Fetch the EIP-1967 implementation account
     #[command(visible_alias = "impl")]
     Implementation {
