+++ crates/common/src/lib.rs
@@ -29,21 +29,10 @@ pub mod shell;
 pub mod term;
 pub mod traits;
 pub mod transactions;
+mod utils;
 
 pub use constants::*;
 pub use contracts::*;
 pub use traits::*;
 pub use transactions::*;
-
-/// Block on a future using the current tokio runtime on the current thread.
-pub fn block_on<F: std::future::Future>(future: F) -> F::Output {
-    block_on_handle(&tokio::runtime::Handle::current(), future)
-}
-
-/// Block on a future using the current tokio runtime on the current thread with the given handle.
-pub fn block_on_handle<F: std::future::Future>(
-    handle: &tokio::runtime::Handle,
-    future: F,
-) -> F::Output {
-    tokio::task::block_in_place(|| handle.block_on(future))
-}
+pub use utils::*;
